Iteration parameter for itable and htable (#4414)
* itable: add iteration index

* hash_table: add iteration index

* itable use (mostly) updated

* hash_table use (mostly) updated

* fix dttools

* fix deltadb

* fix parrot

* fix makeflow

* fix rm

* fix wq

* fix vine

* fix batch_job

* format

* itable: safe remove

* fixes

* fix itable size

* fix compact table

* fix size in clear

* cleanup

* fix safe delete

* fix worker complete tasks loop

* implement safe delete in hash table

* update to new htable

* fix mem leak

* add ro loop for htable

* max worker

* fix release workers

* fix valgrind

* fix wq release workers

* format

* itable_pop efficient

* add hash_table_pop

* use hash_table_pop to release workers

* null guards

* add DELTA_DB_ITERATE

* remove accidental change

* simplify failed_libraries list creation

* correctly initialize ientry

* safe delete for set

* safe iteration for set

* update set usage

* format

+++ dttools/src/itable.c
@@ -5,18 +5,23 @@ This software is distributed under the GNU General Public License.
 See the file COPYING for details.
 */
 
+#include "debug.h"
 #include "itable.h"
 
 #include <stdlib.h>
 #include <string.h>
 
 #define DEFAULT_SIZE 127
-#define DEFAULT_LOAD 0.75
+#define DEFAULT_MAX_LOAD 0.75
+#define DEFAULT_MIN_LOAD 0.125
+
+#define ITERATION_MAX 1000000 /* very conservative overflow protection */
 
 struct entry {
 	UINT64_T key;
 	void *value;
 	struct entry *next;
+	int deleted;
 };
 
 struct itable {
@@ -25,6 +30,8 @@ struct itable {
 	struct entry **buckets;
 	int ibucket;
 	struct entry *ientry;
+	int iteration_index;
+	int need_compact;
 };
 
 struct itable *itable_create(int bucket_count)
@@ -46,7 +53,9 @@ struct itable *itable_create(int bucket_count)
 	}
 
 	h->size = 0;
-
+	h->ientry = 0;
+	h->iteration_index = 0;
+	h->need_compact = 0;
 	return h;
 }
 
@@ -58,7 +67,7 @@ void itable_clear(struct itable *h, void (*delete_func)(void *))
 	for (i = 0; i < h->bucket_count; i++) {
 		e = h->buckets[i];
 		while (e) {
-			if (delete_func)
+			if (delete_func && !e->deleted)
 				delete_func(e->value);
 			f = e->next;
 			free(e);
@@ -68,6 +77,117 @@ void itable_clear(struct itable *h, void (*delete_func)(void *))
 
 	for (i = 0; i < h->bucket_count; i++) {
 		h->buckets[i] = 0;
+	}
+
+	h->size = 0;
+	h->need_compact = 0;
+	h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
+}
+
+static int itable_insert_to_buckets_aux(struct entry **buckets, int bucket_count, struct entry *new_entry)
+{
+	UINT64_T index;
+	struct entry *e;
+
+	index = new_entry->key % bucket_count;
+	e = buckets[index];
+
+	while (e) {
+		/* check that if this key already exist in the table */
+		if (new_entry->key == e->key) {
+			e->value = new_entry->value;
+			e->deleted = 0;
+			free(new_entry);
+			return 1;
+		}
+		e = e->next;
+	}
+
+	new_entry->next = buckets[index];
+	buckets[index] = new_entry;
+
+	return 1;
+}
+
+static int itable_reduce_buckets(struct itable *h)
+{
+	int new_count = ((h->bucket_count + 1) / 2) - 1;
+
+	/* DEFAULT_SIZE is the minimum size */
+	if (new_count < DEFAULT_SIZE) {
+		return 0;
+	}
+
+	/* Table cannot be reduced above DEFAULT_MAX_LOAD */
+	if (((float)h->size / new_count) > DEFAULT_MAX_LOAD) {
+		return 1;
+	}
+
+	struct entry **new_buckets = (struct entry **)calloc(new_count, sizeof(struct entry *));
+	if (!new_buckets) {
+		return 0;
+	}
+
+	struct entry *e, *f;
+	for (int i = 0; i < h->bucket_count; i++) {
+		e = h->buckets[i];
+		while (e) {
+			f = e->next;
+			if (!e->deleted) {
+				e->next = NULL;
+				itable_insert_to_buckets_aux(new_buckets, new_count, e);
+			}
+			e = f;
+		}
+	}
+
+	/* Make the old point to the new */
+	free(h->buckets);
+	h->buckets = new_buckets;
+	h->bucket_count = new_count;
+	h->need_compact = 0;
+
+	h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
+
+	return 1;
+}
+
+/* Compact the table by removing deleted entries. This does not change the size of the table, or the number of buckets. */
+void itable_compact(struct itable *h)
+{
+	struct entry *e, *f;
+	int i;
+
+	if (!h->need_compact) {
+		return;
+	}
+
+	for (i = 0; i < h->bucket_count; i++) {
+		e = h->buckets[i];
+
+		/* find the head of the bucket and remove deleted entries*/
+		while (e && e->deleted) {
+			f = e->next;
+			free(e);
+			e = f;
+		}
+		h->buckets[i] = e;
+
+		/* remove deleted entries from the rest of the bucket */
+		while (e) {
+			f = e->next;
+			while (f && f->deleted) {
+				e->next = f->next;
+				free(f);
+				f = e->next;
+			}
+			e = f;
+		}
+	}
+
+	h->need_compact = 0;
+	if (((float)h->size / h->bucket_count) < DEFAULT_MIN_LOAD) {
+		itable_reduce_buckets(h);
 	}
 }
 
@@ -93,7 +213,10 @@ void *itable_lookup(struct itable *h, UINT64_T key)
 
 	while (e) {
 		if (key == e->key) {
-			return e->value;
+			if (!e->deleted) {
+				return e->value;
+			}
+			return 0;
 		}
 		e = e->next;
 	}
@@ -103,41 +226,32 @@ void *itable_lookup(struct itable *h, UINT64_T key)
 
 static int itable_double_buckets(struct itable *h)
 {
-	struct itable *hn = itable_create(2 * h->bucket_count);
-
-	if (!hn)
+	int new_count = (2 * (h->bucket_count + 1)) - 1;
+	struct entry **new_buckets = (struct entry **)calloc(new_count, sizeof(struct entry *));
+	if (!new_buckets) {
 		return 0;
+	}
 
-	/* Move pairs to new hash */
-	uint64_t key;
-	void *value;
-	itable_firstkey(h);
-	while (itable_nextkey(h, &key, &value))
-		if (!itable_insert(hn, key, value)) {
-			itable_delete(hn);
-			return 0;
-		}
-
-	/* Delete all old pairs */
 	struct entry *e, *f;
-	int i;
-	for (i = 0; i < h->bucket_count; i++) {
+	for (int i = 0; i < h->bucket_count; i++) {
 		e = h->buckets[i];
 		while (e) {
 			f = e->next;
-			free(e);
+			if (!e->deleted) {
+				e->next = NULL;
+				itable_insert_to_buckets_aux(new_buckets, new_count, e);
+			}
 			e = f;
 		}
 	}
 
 	/* Make the old point to the new */
 	free(h->buckets);
-	h->buckets = hn->buckets;
-	h->bucket_count = hn->bucket_count;
-	h->size = hn->size;
+	h->buckets = new_buckets;
+	h->bucket_count = new_count;
+	h->need_compact = 0;
 
-	/* Delete reference to new, so old is safe */
-	free(hn);
+	h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
 
 	return 1;
 }
@@ -147,7 +261,7 @@ int itable_insert(struct itable *h, UINT64_T key, const void *value)
 	struct entry *e;
 	UINT64_T index;
 
-	if (((float)h->size / h->bucket_count) > DEFAULT_LOAD)
+	if (((float)h->size / h->bucket_count) > DEFAULT_MAX_LOAD)
 		itable_double_buckets(h);
 
 	index = key % h->bucket_count;
@@ -155,7 +269,13 @@ int itable_insert(struct itable *h, UINT64_T key, const void *value)
 
 	while (e) {
 		if (key == e->key) {
+			int was_deleted = e->deleted;
 			e->value = (void *)value;
+			e->deleted = 0;
+			if (was_deleted) {
+				h->size++;
+			}
+			h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
 			return 1;
 		}
 		e = e->next;
@@ -167,36 +287,43 @@ int itable_insert(struct itable *h, UINT64_T key, const void *value)
 
 	e->key = key;
 	e->value = (void *)value;
-	e->next = h->buckets[index];
-	h->buckets[index] = e;
-	h->size++;
+	e->deleted = 0;
+
+	int inserted = itable_insert_to_buckets_aux(h->buckets, h->bucket_count, e);
+	if (inserted) {
+		h->size++;
+		h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
+	} else {
+		free(e);
+		return 0;
+	}
 
 	return 1;
 }
 
+/* keys are not really removed, just marked as deleted. A call to itable_compact will remove them. */
 void *itable_remove(struct itable *h, UINT64_T key)
 {
-	struct entry *e, *f;
+	struct entry *e;
 	void *value;
 	UINT64_T index;
 
 	index = key % h->bucket_count;
 	e = h->buckets[index];
-	f = 0;
 
 	while (e) {
 		if (key == e->key) {
-			if (f) {
-				f->next = e->next;
-			} else {
-				h->buckets[index] = e->next;
+			if (e->deleted) {
+				return 0;
 			}
+			e->deleted = 1;
+			h->need_compact = 1;
 			value = e->value;
-			free(e);
+
 			h->size--;
 			return value;
 		}
-		f = e;
+
 		e = e->next;
 	}
 
@@ -208,37 +335,65 @@ void *itable_pop(struct itable *t)
 	UINT64_T key;
 	void *value;
 
-	itable_firstkey(t);
-	if (itable_nextkey(t, &key, (void **)&value)) {
-		return itable_remove(t, key);
-	} else {
-		return 0;
+	if (!t) {
+		return NULL;
 	}
+
+	if (itable_nextkey(t, t->iteration_index, &key, (void **)&value)) {
+		t->iteration_index++;
+		return itable_remove(t, key);
+	}
+
+	itable_firstkey(t);
+
+	if (itable_nextkey(t, t->iteration_index, &key, (void **)&value)) {
+		return itable_remove(t, key);
+	}
+
+	return NULL;
 }
 
-void itable_firstkey(struct itable *h)
+int itable_firstkey(struct itable *h)
 {
+	itable_compact(h);
+	h->iteration_index = (h->iteration_index + 1) % ITERATION_MAX;
+
 	h->ientry = 0;
 	for (h->ibucket = 0; h->ibucket < h->bucket_count; h->ibucket++) {
 		h->ientry = h->buckets[h->ibucket];
-		if (h->ientry)
+		if (h->ientry) {
+			/* we don't need to check for deleted entries here because itable_compact removes them */
 			break;
+		}
 	}
+
+	return h->iteration_index;
 }
 
-int itable_nextkey(struct itable *h, UINT64_T *key, void **value)
+int itable_nextkey(struct itable *h, int iteration, UINT64_T *key, void **value)
 {
+	if (!h) {
+		return 0;
+	}
+
+	if (iteration != h->iteration_index) {
+		fatal("cctools bug: the itable iteration has not been reset since last modification");
+	}
+
 	if (h->ientry) {
 		*key = h->ientry->key;
 		if (value)
 			*value = h->ientry->value;
 
 		h->ientry = h->ientry->next;
+		while (h->ientry && h->ientry->deleted) {
+			h->ientry = h->ientry->next;
+		}
 		if (!h->ientry) {
 			h->ibucket++;
 			for (; h->ibucket < h->bucket_count; h->ibucket++) {
 				h->ientry = h->buckets[h->ibucket];
-				if (h->ientry)
+				if (h->ientry && !h->ientry->deleted)
 					break;
 			}
 		}
@@ -246,6 +401,28 @@ int itable_nextkey(struct itable *h, UINT64_T *key, void **value)
 		return 1;
 	} else {
 		return 0;
+	}
+}
+
+void itable_foreach_ro(struct itable *h, void (*func)(UINT64_T key, void *value, void *arg), void *arg)
+{
+	struct entry *e;
+	int i;
+
+	int prev_iteration = h->iteration_index;
+
+	for (i = 0; i < h->bucket_count; i++) {
+		e = h->buckets[i];
+		while (e) {
+			if (!e->deleted) {
+				func(e->key, e->value, arg);
+			}
+			e = e->next;
+
+			if (h->iteration_index != prev_iteration) {
+				fatal("cctools bug: the itable iteration has been modified since last call to itable_foreach_ro");
+			}
+		}
 	}
 }
